Move match function on the Server model.
